requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cssselect2==0.8.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==11.3.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
